get_metadata refactored
