+++ src/sr2silo/vpipe/metadata.py
@@ -2,6 +2,13 @@
 
 from __future__ import annotations
 
+import csv
+import datetime
+import logging
+from pathlib import Path
+
+import yaml
+
 
 def sample_id_decoder(sample_id: str) -> dict:
     """Decode the sample ID into individual components.
@@ -50,3 +57,133 @@ def batch_id_decoder(batch_id: str) -> dict:
         "sequencing_date": sequencing_date,
         "flow_cell_serial_number": flow_cell_serial_number,
     }
+
+
+def convert_to_iso_date(date: str) -> str:
+    """Convert a date string to ISO 8601 format (date only)."""
+    # Parse the date string
+    date_obj = datetime.datetime.strptime(date, "%Y-%m-%d")
+    # Format the date as ISO 8601 (date only)
+    return date_obj.date().isoformat()
+
+
+def enrich_metadata_from_timeline(metadata: dict, timeline: Path) -> None:
+    """Enrich metadata from the timeline file."""
+    if not timeline.is_file():
+        logging.error(f"Timeline file not found or is not a file: {timeline}")
+        raise FileNotFoundError(f"Timeline file not found or is not a file: {timeline}")
+
+    with timeline.open() as f:
+        reader = csv.reader(f, delimiter="\t")
+        for row in reader:
+            sample_id_match = row[0] == metadata["sample_id"]
+            batch_id_match = row[1] == metadata["batch_id"]
+
+            if sample_id_match and batch_id_match:
+                logging.info(
+                    "Enriching metadata with timeline data e.g. read_length, "
+                    "primer_protocol, location_name"
+                )
+                metadata["read_length"] = row[2]
+                metadata["primer_protocol"] = row[3]
+                metadata["location_name"] = row[6]
+
+                # Convert sampling_date to ISO format for comparison
+                timeline_sampling_date = convert_to_iso_date(row[5])
+
+                location_code_mismatch = int(metadata["location_code"]) != int(row[4])
+                sampling_date_mismatch = (
+                    metadata["sampling_date"] != timeline_sampling_date
+                )
+
+                if location_code_mismatch:
+                    logging.warning(
+                        f"Mismatch in location code for sample_id "
+                        f"{metadata['sample_id']} and batch_id "
+                        f"{metadata['batch_id']}"
+                    )
+                    logging.debug(
+                        f"Location code mismatch: {metadata['location_code']} "
+                        f"(sample_id) vs {row[4]} (timeline)"
+                    )
+                    logging.debug(
+                        f"Location code types: {type(metadata['location_code'])} "
+                        f"(sample_id) vs {type(row[4])} (timeline)"
+                    )
+
+                if sampling_date_mismatch:
+                    logging.warning(
+                        f"Mismatch in sampling date for sample_id "
+                        f"{metadata['sample_id']} and batch_id "
+                        f"{metadata['batch_id']}"
+                    )
+                    logging.debug(
+                        f"Sampling date mismatch: {metadata['sampling_date']} "
+                        f"(sample_id) vs {timeline_sampling_date} (timeline)"
+                    )
+                    logging.debug(
+                        f"Sampling date types: {type(metadata['sampling_date'])} "
+                        f"(sample_id) vs {type(timeline_sampling_date)} (timeline)"
+                    )
+                break
+        else:
+            raise ValueError(
+                f"No matching entry found in timeline for sample_id "
+                f"{metadata['sample_id']} and batch_id {metadata['batch_id']}"
+            )
+
+
+def get_metadata(sample_id: str, batch_id: str, timeline: Path, primers: Path) -> dict:
+    """
+    Get metadata for a given sample and batch directory.
+    Cross-references the directory with the timeline file to get the metadata.
+
+    Args:
+        sample_id (str): The sample ID to use for metadata.
+        batch_id (str): The batch ID to use for metadata.
+        timeline (Path): The timeline file to cross-reference the metadata.
+        primers (Path): The primers file to cross-reference the metadata.
+
+    Returns:
+        dict: A dictionary containing the metadata.
+
+    """
+
+    metadata = {}
+    metadata["sample_id"] = sample_id
+    metadata["batch_id"] = batch_id
+
+    # Decompose the ids into individual components
+    logging.info(f"Decoding sample_id: {metadata['sample_id']}")
+    sample_id = metadata["sample_id"]
+    metadata.update(sample_id_decoder(sample_id))
+    logging.info(f"Decoding batch_id: {metadata['batch_id']}")
+    batch_id = metadata["batch_id"]
+    metadata.update(batch_id_decoder(batch_id))
+
+    enrich_metadata_from_timeline(metadata, timeline)
+
+    # Read the primers yaml to get additional metadata
+    # find the key with matching primer_protocol and get the "name" value
+    # as the canonical name of the primer protocol
+    if not primers.is_file():
+        logging.error(f"Primers file not found or is not a file: {primers}")
+        raise FileNotFoundError(f"Primers file not found or is not a file: {primers}")
+    # Load YAML file
+    with open(primers, "r") as file:
+        primers_conf = yaml.safe_load(file)
+    logging.debug(f"Primers: {primers_conf}")
+    logging.debug(f" Type of primers: {type(primers_conf)}")
+    for primer in primers_conf.keys():
+        if primer == metadata["primer_protocol"]:
+            logging.info(
+                "Enriching metadata with primer data e.g. primer_protocol_name"
+            )
+            metadata["primer_protocol_name"] = primers_conf[primer]["name"]
+            break
+    else:
+        raise ValueError(
+            f"No matching entry found in primers for primer_protocol"
+            f"{metadata['primer_protocol']}"
+        )
+    return metadata
